UI Menu Scale Implemented.
Menu_Scale class implemented:
- Has a min/max, interval step, and starting value;
- Left arrow will reduce the current value by interval step unless at minimum;
- Right arrow will add the interval to the current val unless at maximum;
- Left and Right arrows are ZEngine::Buttons;
- Re-organised file structure so that scale buttons and menu buttons each have their own folders for their respective .png files and .ttf files.
- Made basic scale button images.

Updated button class so that all of the input handling (pressing/releasing) is done internally on the buttons, not via the InputManager.

Also updated button class to get the textures/fonts internally, not via the class they are a part of.
Working to scrap the Input Manager altogether.