Shop UI and Functionality
Used the Menu Scale class to implement shop UI and functionality.
Shop is a gamestate that is added on top of the gameplay state, pulls the gun stats and dumps them into a bunch of MenuScales.
Also added in Zombits (Zb) as a currency/tracking for them.

Note: Also began to organise the classes and code into filters, as well as organised the media files structure.