Shop Gun Scales now are tied to costs
Created a ShopGunScale class that extends the Menu_Scale class.
It's effectively the same with the addition of a cost tied to the scale. Can't upgrade without enough zbits.

TODO
Add a zbits text object to the shop menu.
Make sure that gun stats and the menu numbers are persistent.
Maybe make the shop gamestate pause when in gameplay? That way it's not creating a new shop menu each time.