Debug Drawing, Individual tile loading;
Tiles can be loaded in from a tilesheet;
Tiles can be defined as collidable, which sets a collision box for them;

Also fixed the collision box for the player sprite.

Implemented a debug-draw mechanic
Press O to enter debug mode and show all of the physics object and their boundaries.