Level & Level Builder framework;
The foundations for a level (loading and saving) system is put in place with two new classes, Room for individual rooms, and Level, for the collection of rooms.

Additionally completed the basic UI foundation for the level builder.

TODO:
Add in room-saving and room-loading for level construction, consider design for how the level actually maps out the rooms.