Level-Building, saving levels, and loading levels;
When L is pressed in-game, the game enters a "building state" that lets the user build a level by clicking to replace tiles in game. The user can scroll to change which tile is selected.

Pressing M enters a save state, where the user then types in the name of the level they'd like to save, pressing N enters a load state where the user types in the name of the level they'd like to load.

Levels are saved as .txt files in the games LevelData folder. 

Additionally, jumping is now a standardised height.