Tile picker implemented. Camera panning added.
When in the level builder use the arrow keys to pan the camera around the level.

In the level builder, right click to open the tile selector. 

Left click on a tile to close the selector and pick a tile to build with. 

The current active tile will have a yellow outline for easy re-access on an accidental click.