Level Editor updated;
Level editor now has proper pages where Units/Obstacles can be added to it whenever they are added in via new code.

The new pages (2 for units, 3 for obstacles) function the same as the first page.

It still needs to be updated for props, but that can wait until the more pressing items are finished.