It's A* me, Pathfinding! (Oh, Baldy too)
Added in a basic implementation of Baldy, a basic pirate enemy.
All of Baldy's animations are loaded in and prepared.


Summarily, Pathfinding is implemented via the A* algorithm. 
Node class implemented, pathfinding process is TEMPORARILY implemented on Baldy, will move the pathfinding implementation to either a new class or to the Utilities class on the ZEngine.

Nodes are automatically generated on the tile map in the spaces above collidable tiles.

After the nodes in a map are generated, ALL of the nodes generate links between the node and the other nodes in the level with viable links.

Links are stored as an 'Edge', a struct which contains a pointer to a node (the node that this node is connected to) and a 'cost' (the cost of the edge as it relates to pathfinding.

A* is then run on the list of nodes.


NOTES:
Node generation isn't perfect, I need to do work on costing the nodes a bit better.