A* with jumping BAY BEEEEEE
A* is working with Baldy AND IT JUMPSSSSSSS

Work needed to balance it out and manage the following issues:

Baldy might need to regenerate it's path when he gets stuck.

Should consider moving the pathfinding generation code to NavMap and maintain that the AI just commits to asking that map for a path using it's maximum jump power and forward velocity (to stop it from jumping when it shouldn't)

Definitely need to prioritise updating the costing functionality to calculate the length of the jump trajectory so that we prio the smallest jumps when available.

Also need to update the jump map generation so that it takes into account the size of baldy - lets me set up the navmapping for other AI easily!